prime b

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -3,14 +3,14 @@
 /**
  * main -  calculate largest prime of 612852475143
  *
- * Return: Success Always
+ * Return: 0 Always
  */
 
 
 int main(void)
 {
-int number = 612852475143;
-int p;
+long int number = 612852475143;
+long int p;
 for (p = 2; p < number; p++)
 {
 if (number % p == 0)
